add useHover
