+++ src/helpers/index.ts
@@ -1,6 +1,8 @@
 export { default as active } from './active';
 export * from './groupBy';
-export * from './Hover';
+export { default as Hover } from './Hover';
+export * from './HoverStyle';
 export { default as icons } from './icons';
 export * from './slack';
 export * from './strings';
+export * from './useHover';
