+++ src/helpers/useHover.ts
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const HoverContext = React.createContext(false);
+
+export const useHover = () => React.useContext(HoverContext);
